Update AppContext
Add Contact Modal control functionality

Migrate utilities into app directory
